mytime/requirements.txt
click==8.1.7
numpy==1.26.1
pandas==2.1.1
pendulum==2.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
pytzdata==2020.1
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
